Adds dudetray - a GTK+ system tray that monitors git-dude notifications
git-dude is updated to write to disk files in ~/.dude-notifications when a new
update to a watched repo is processed. The dudetray monitors this file folder
and stores the notifications in it's tray. Right clicking the tray icon will
display a list of notifications that have come through.

To install, cd into the dudetray folder and run make. This will create the
dudetray binary which can be placed anywhere on your machine.

The icon image is compiled into the binary using GTK+ inline pixbuf storage.
This has only been tested on GNU/Linux Ubuntu 12.10 (64-bit and 32-bit).